Hamburger Menu  720p Ready New Font Bio Page
I added a hamburger menu to connect people to the team and bio pages
soon to be added. Also I want to add a killer feature that would make
these articles automatically convert to spanish. Changed the font
because my groupmates believed brushed was unprofessional lastly I made
the homepage and article page 720p ready meaning chromebook ready so
when these articles are viewed on chromebooks they look the same as on
desktop or 1080p screens.